Extract diff from PR 36171
