+++ src/Common/FileSegment.cpp
@@ -1,10 +1,11 @@
 #include "FileSegment.h"
 #include <base/getThreadId.h>
-#include <Common/FileCache.h>
 #include <Common/hex.h>
+#include <Common/logger_useful.h>
 #include <IO/WriteBufferFromString.h>
 #include <IO/Operators.h>
 #include <filesystem>
+
 
 namespace CurrentMetrics
 {
@@ -25,7 +26,8 @@ FileSegment::FileSegment(
         size_t size_,
         const Key & key_,
         IFileCache * cache_,
-        State download_state_)
+        State download_state_,
+        bool is_persistent_)
     : segment_range(offset_, offset_ + size_ - 1)
     , download_state(download_state_)
     , file_key(key_)
@@ -35,6 +37,7 @@ FileSegment::FileSegment(
 #else
     , log(&Poco::Logger::get("FileSegment"))
 #endif
+    , is_persistent(is_persistent_) /// Not really used for now, see PR 36171
 {
     /// On creation, file segment state can be EMPTY, DOWNLOADED, DOWNLOADING.
     switch (download_state)
@@ -241,7 +244,7 @@ void FileSegment::write(const char * from, size_t size, size_t offset_)
                             "Cache writer was finalized (downloaded size: {}, state: {})",
                             downloaded_size, stateToString(download_state));
 
-        auto download_path = cache->getPathInLocalCache(key(), offset());
+        auto download_path = getPathInLocalCache();
         cache_writer = std::make_unique<WriteBufferFromFile>(download_path);
     }
 
@@ -271,6 +274,11 @@ void FileSegment::write(const char * from, size_t size, size_t offset_)
     assert(getDownloadOffset() == offset_ + size);
 }
 
+String FileSegment::getPathInLocalCache() const
+{
+    return cache->getPathInLocalCache(key(), offset(), isPersistent());
+}
+
 void FileSegment::writeInMemory(const char * from, size_t size)
 {
     if (!size)
@@ -287,7 +295,7 @@ void FileSegment::writeInMemory(const char * from, size_t size)
     if (cache_writer)
         throw Exception(ErrorCodes::LOGICAL_ERROR, "Cache writer already initialized");
 
-    auto download_path = cache->getPathInLocalCache(key(), offset());
+    auto download_path = getPathInLocalCache();
     cache_writer = std::make_unique<WriteBufferFromFile>(download_path, size + 1);
 
     try
@@ -677,7 +685,7 @@ void FileSegment::assertCorrectnessImpl(std::lock_guard<std::mutex> & /* segment
 {
     assert(downloader_id.empty() == (download_state != FileSegment::State::DOWNLOADING));
     assert(!downloader_id.empty() == (download_state == FileSegment::State::DOWNLOADING));
-    assert(download_state != FileSegment::State::DOWNLOADED || std::filesystem::file_size(cache->getPathInLocalCache(key(), offset())) > 0);
+    assert(download_state != FileSegment::State::DOWNLOADED || std::filesystem::file_size(getPathInLocalCache()) > 0);
 }
 
 void FileSegment::throwIfDetached() const
@@ -729,6 +737,7 @@ FileSegmentPtr FileSegment::getSnapshot(const FileSegmentPtr & file_segment, std
     snapshot->ref_count = file_segment.use_count();
     snapshot->downloaded_size = file_segment->getDownloadedSize();
     snapshot->download_state = file_segment->state();
+    snapshot->is_persistent = file_segment->isPersistent();
 
     return snapshot;
 }
