+++ src/Disks/IO/CachedReadBufferFromRemoteFS.h
@@ -1,11 +1,12 @@
 #pragma once
 
-#include <Common/FileCache.h>
+#include <Common/IFileCache.h>
 #include <IO/SeekableReadBuffer.h>
 #include <IO/WriteBufferFromFile.h>
 #include <IO/ReadSettings.h>
 #include <Common/logger_useful.h>
 #include <Interpreters/FilesystemCacheLog.h>
+#include <Common/FileSegment.h>
 
 
 namespace CurrentMetrics
@@ -125,6 +126,8 @@ private:
     ProfileEvents::Counters current_file_segment_counters;
 
     IFileCache::QueryContextHolder query_context_holder;
+
+    bool is_persistent;
 };
 
 }
