+++ src/Interpreters/AsynchronousMetrics.cpp
@@ -10,7 +10,7 @@
 #include <Common/typeid_cast.h>
 #include <Common/filesystemHelpers.h>
 #include <Common/FileCacheFactory.h>
-#include <Common/FileCache.h>
+#include <Common/IFileCache.h>
 #include <Server/ProtocolServerAdapter.h>
 #include <Storages/MarkCache.h>
 #include <Storages/StorageMergeTree.h>
